+++ module/utilities/handlebars.js
@@ -1,3 +1,5 @@
+import {byNameAscending} from "./sorting.js";
+
 export async function registerHandlebars() {
   Handlebars.registerHelper("expanded", function (itemID) {
     if (game.user.expanded != undefined) {
@@ -100,5 +102,54 @@ export async function registerHandlebars() {
     }
 
     return output;
+  });
+
+  Handlebars.registerHelper("tagOnItem", function (array, value) {
+    if (!Array.isArray(array)) array = [];
+    if (!value) value = "";
+    if (array.includes(value)) {
+      return "";
+    } else {
+      return "tag-inactive";
+    }
+  });
+
+  Handlebars.registerHelper("activeTags", function (actorSheet, tagIDs, recursionIDs) {
+    if (!Array.isArray(tagIDs)) tagIDs = [];
+    if (!Array.isArray(recursionIDs)) recursionIDs = [];
+    let tagArray = [];
+    let recursionArray = [];
+    let tagOutput = "";
+    let recursionOutput = "";
+    let isObserver = (actorSheet.sheetSettings.isObserver) ? "disabled" : "";
+
+    for (let item of actorSheet.items) {
+      if (item.type == "tag" && tagIDs.includes(item._id) && actorSheet.actor.system.settings.general.tags.active) {
+        tagArray.push(item);
+      } else if (item.type == "recursion" && recursionIDs.includes(item._id) && actorSheet.actor.system.settings.general.gameMode == "Strange") {
+        recursionArray.push(item);
+      }
+    }
+
+    tagArray.sort(byNameAscending);
+    recursionArray.sort(byNameAscending);
+
+    for (let tag of tagArray) {
+      let inactive = (!tag.system.active) ? "tag-inactive" : "";
+      let exclusive = (tag.system.exclusive) ? "<i class='fas fa-exclamation'></i>" : "";
+      let title = (tag.system.active) ? game.i18n.format("CYPHERSYSTEM.ArchiveItemsWithTag", {tag: tag.name}) : game.i18n.format("CYPHERSYSTEM.UnarchiveItemsWithTag", {tag: tag.name});
+      tagOutput = tagOutput + `<a class="tag-items toggle-tag ${inactive} ${isObserver}" data-item-id="${tag._id}" title="${title}">` + exclusive + `<i class="fas fa-hashtag"></i> ${tag.name}</a> `;
+    }
+
+    for (let recursion of recursionArray) {
+      let inactive = (!recursion.system.active) ? "tag-inactive" : "";
+      let title = game.i18n.format("CYPHERSYSTEM.TranslateToRecursion", {recursion: recursion.name});
+      recursionOutput = recursionOutput + `<a class="tag-items toggle-tag ${inactive} ${isObserver}" data-item-id="${recursion._id}" title="${title}"><i class="fas fa-at"></i> ${recursion.name}</a> `;
+    }
+
+    if (tagOutput) tagOutput = "<p>" + tagOutput + "</p>";
+    if (recursionOutput) recursionOutput = "<p>" + recursionOutput + "</p>";
+
+    return recursionOutput + tagOutput;
   });
 }